fix problem while updating and update components
